wrote tenant service and repo for postgresql. Implemented get all function
