+++ pkg/tenant/delivery/http/tenant.go
@@ -1,0 +1,1 @@
+package http
